AI_Assistant/frontend/requirements.txt
streamlit>=1.22.0
requests>=2.28.0
pandas>=1.5.0
openpyxl>=3.0.10
python-dotenv>=1.0.0
